Two new additions: Key cards determine the color id of a generated deck. Also, you can just specify the color id directly.

+++ deck_builder/magique/interactions.h
@@ -15,7 +15,7 @@ class interactions
 public:
     interactions(std::string path);
 
-    double evaluate(const card &a, const card &b) const;
+    uint64_t evaluate(const card &a, const card &b) const;
 private:
     std::array<std::array<double, 250>, 250> interactions_store_; // TODO hard-coded number of classes, ugh.
 };
